requirements.txt
flask
gunicorn
psycopg2-binary
reportlab
arabic-reshaper
python-bidi



